change byte to int in studentDto

+++ ExamApp.Services/DTOs/StudentDto.cs
@@ -2,9 +2,9 @@
 {
     public class StudentDto
     {
-        public required int StudentNumber { get; init; } 
-        public required string Firstname { get; init; } = string.Empty; 
-        public required string Lastname { get; init; } = string.Empty; 
-        public required byte ClassNumber { get; init; }  
+        public required int StudentNumber { get; init; }
+        public required string Firstname { get; init; } = string.Empty;
+        public required string Lastname { get; init; } = string.Empty;
+        public required int ClassNumber { get; init; }   // чтобы ловить негативные значения при валидации в сущности это поле байт
     }
 }
